+++ game/king.py
@@ -10,10 +10,10 @@ class King(Piece):
         return " ♚" if self.__color__ == "WHITE" else " ♔"
         
     def permited_move(self, from_row, from_col, to_row, to_col, board):
-        if abs(from_row - to_row) <= 1 and abs(from_col - to_col) <= 1 and not (from_row == to_row and from_col == to_col):
-            return True
-        else: 
-            return False
+        if abs(from_row - to_row) <= 1 and abs(from_col - to_col) <= 1:
+            if self.permited_move_diagonal(from_row, from_col, to_row, to_col, board) or self.permited_move_orthogonal(from_row, from_col, to_row, to_col, board):
+                return True
+        return False