+++ game/pawn.py
@@ -23,19 +23,21 @@ class Pawn(Piece):
         destination_piece = board.get_piece(to_row, to_col)
         return destination_piece != "No piece" and destination_piece[1] != self.__color__
 
-
     def permited_move(self, from_row, from_col, to_row, to_col, board):
         direction = -1 if self.__color__ == "WHITE" else 1
-        if to_col == from_col:  
-            if (to_row - from_row) == direction and board.get_piece(to_row, to_col) == "No piece":
+        
+        # Movimiento ortogonal (avance)
+        if self.permited_move_orthogonal(from_row, from_col, to_row, to_col, board):
+            if (to_row - from_row) == direction and self.empty_square(board, to_row, to_col):
                 return True
-            
             if self.is_valid_starting_row(from_row) and self.can_do_double_step(from_row, to_row, direction) and self.empty_square(board, to_row, to_col):
                 return True
-                
-        if abs(to_col - from_col) == 1 and (to_row - from_row) == direction:
-            if self.is_enemy_piece(board, to_row, to_col):
-                return True        
+        
+        # Movimiento diagonal (captura)
+        if self.permited_move_diagonal(from_row, from_col, to_row, to_col, board):
+            if (to_row - from_row) == direction and self.is_enemy_piece(board, to_row, to_col):
+                return True
+        
+        return False
 
-        return False