+++ game/queen.py
@@ -9,13 +9,11 @@ class Queen(Piece):
         return " ♛" if self.__color__ == "WHITE" else " ♕"
 
     def permited_move(self, from_row, from_col, to_row, to_col, board):
-        if to_row == from_row and to_col != from_col:
-            return True
-        elif to_col == from_col and to_row != from_row:
-            return True
-        elif abs(to_row - from_row) == abs(to_col - from_col): 
+        permited_move_diag = self.permited_move_diagonal(from_row, from_col, to_row, to_col, board)
+        permitted_move_orthogonal = self.permited_move_orthogonal(from_row, from_col, to_row, to_col, board)
+        if permited_move_diag or permitted_move_orthogonal == True:
             return True
         else:
             return False
-        
+