+++ game/rook.py
@@ -1,6 +1,6 @@
-from game.piece import Piece
+from game.piece import Piece as piece
 
-class Rook(Piece):
+class Rook(piece):
     def __init__(self, color):
         super().__init__(color)
         self.__type__ = "ROOK"
@@ -8,11 +8,5 @@ class Rook(Piece):
     def __str__(self):
         return " ♜" if self.__color__ == "WHITE" else " ♖"
 
- 
     def permited_move(self, from_row, from_col, to_row, to_col, board):
-        if to_row == from_row and to_col != from_col:
-            return True
-        elif to_col == from_col and to_row != from_row:
-            return True
-        return False
-    
+        return self.permited_move_orthogonal(from_row, from_col, to_row, to_col, board)